Move `Directory` to P_CacheDir.h
